% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_annotation.R
\name{add_canopus}
\alias{add_canopus}
\title{Add CANOPUS classification}
\usage{
add_canopus(annotation_table, molclass_table, canopus_summary_file)
}
\arguments{
\item{annotation_table}{Dataframe with annotations}
}
\value{
Dataframe with molecular classification
}
\description{
Function to add molecular classes predicted by CANOPUS
}
