% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_alignment.R
\name{apply_alignment}
\alias{apply_alignment}
\alias{apply_alignment.pg}
\alias{apply_alignment.ow}
\alias{apply_alignment.lama}
\title{Apply Peak alignment}
\usage{
apply_alignment(data, metadata, method = c("pg", "ow", "lama"), ...)

apply_alignment.pg(
  data,
  metadata,
  group_by,
  bin_size = 0.25,
  ppm_bin = 0,
  min_fraction = 0.9,
  extra_peaks = 1,
  smooth = c("loess", "linear"),
  span = 1,
  family = c("gaussian", "symmetric"),
  subset_samples = NULL,
  subset_adjust = c("previous", "average"),
  ...
)

apply_alignment.ow(
  data,
  metadata,
  bin_size = 1,
  smooth_responsiveness = 1,
  distance_function = c("cor", "cor_opt", "cov", "prd", "euc"),
  local_alignment = FALSE,
  subset_samples = NULL,
  subset_adjust = c("previous", "average"),
  ...
)

apply_alignment.lama(
  data,
  metadata,
  lama_file,
  lama_method = c("loess", "gam"),
  span_lama = 0.5,
  ppm = 20,
  tolerance_mz = 0,
  tolerance_rt = 5,
  gam_smoothing = c("tp", "ts", "ds", "cr", "cs", "cc", "sos", "bs", "ps", "re", "mrf",
    "gp", "so"),
  ...
)
}
\arguments{
\item{data}{An \linkS4class{MSnExp} object in \emph{centroid} mode.}

\item{metadata}{Sample information data.frame.}
}
\value{
MSnExp-class object after alignment.
}
\description{
Function to apply peak alignment
}
