% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Module_spectraAlignment.R
\name{alignSpectraUI}
\alias{alignSpectraUI}
\title{Align spectra UI}
\usage{
alignSpectraUI(id)
}
\arguments{
\item{id}{character used to specify namespace, see \code{\link[shiny:NS]{shiny::NS}}}
}
\value{
a \code{shiny::\link[shiny]{tagList}} containing UI elements
}
\description{
Align spectra UI
}
