% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_functions.R
\name{annotate_CAMERA}
\alias{annotate_CAMERA}
\title{Adduct annotation with CAMERA}
\usage{
annotate_CAMERA(data, group_by)
}
\arguments{
\item{data}{An \linkS4class{MSnExp} object with peaks picked}

\item{group_by}{column}
}
\description{
Function to apply peak picking parameters on a data subset
}
