% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_annotation_SIRIUS.R
\name{annotate_SIRIUS}
\alias{annotate_SIRIUS}
\title{Annotate MS2 using SIRIUS}
\usage{
annotate_SIRIUS(ms2.data, groupdir = "grouped_mgf", outdir = "SIRIUS_output")
}
\arguments{
\item{ms2.data}{An \link{MSpectra} object containing MS2 data}

\item{ms1.data}{A \link{data.frame} containing the columns 'FeatureID', 'mzmed' and 'rtmed'}
}
\value{
A dataframe with the results of annotation using SIRIUS
}
\description{
Function to identify feature structure using SIRIUS
(https://bio.informatik.uni-jena.de/software/sirius/)
}
