% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UI_annotation_app.R
\name{annotationappUI}
\alias{annotationappUI}
\title{Solo annotation UI
`}
\usage{
annotationappUI(id)
}
\arguments{
\item{id}{character used to specify namespace, see \code{\link[shiny:NS]{shiny::NS}}}
}
\value{
a \code{shiny::\link[shiny]{tagList}} containing UI elements
}
\description{
Solo annotation UI
`
}
