% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing_functions.R
\name{apply_alignment}
\alias{apply_alignment}
\title{Apply Peak alignment}
\usage{
apply_alignment(
  data,
  metadata,
  min_frac,
  min_samples,
  bin = 0.25,
  bw = 30,
  group_by,
  plot = FALSE
)
}
\arguments{
\item{data}{An \linkS4class{MSnExp} object in \emph{centroid} mode.}

\item{metadata}{Sample information data.frame.}

\item{min_frac}{Minimum fraction of samples within a sample group a peak must
be present to be considered a peak group.}

\item{min_samples}{Minimum number of samples in at least one sample group a
peak must be present to be considered a peak group.}

\item{bin}{Size of the overlapping \emph{mz} slices}

\item{bw}{= Bandwidth of the smoothing kernel}

\item{group_by}{Vector with the same length of the samples with grouping
information}
}
\description{
Function to apply peak picking parameters on a data subset
}
