% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing_functions.R
\name{apply_gap_filling}
\alias{apply_gap_filling}
\title{Apply Gap Filling}
\usage{
apply_gap_filling(data)
}
\arguments{
\item{data}{An \linkS4class{MSnExp} object in \emph{centroid} mode.}
}
\description{
Function to apply peak picking parameters on a data subset
}
