% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_preprocessing.R
\name{apply_peak_picking}
\alias{apply_peak_picking}
\title{Apply Peak Picking}
\usage{
apply_peak_picking(
  data,
  method,
  ppm = 25,
  p_width = c(20, 50),
  snt = 3,
  noise = 1e+06,
  prefilter = c(1, 100),
  mz_diff = 0.001,
  bin = 0.1,
  fwhm = 30,
  sigma = 12.72,
  max = 10,
  steps = 2
)
}
\arguments{
\item{data}{An \linkS4class{MSnExp} object in \emph{centroid} mode.}

\item{p_width}{Minium and maximum allowed peak width.}

\item{snt}{Minimum signal-to-noise threshold allowed.}

\item{noise}{Noise threshold.}

\item{prefilter}{Prefilter step cutoff (\code{c(k, I)}). Mass will be retained if
they contain at least \code{k} peaks with intensity >= \code{I}.}
}
\description{
Function to apply peak picking parameters on a data subset
}
