% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_preprocessing.R
\name{apply_peak_refinement}
\alias{apply_peak_refinement}
\title{Apply Peak Refinement}
\usage{
apply_peak_refinement(
  data,
  metadata,
  expand_rt = 2,
  expand_mz = 0,
  ppm = 10,
  min_prop = 0.75
)
}
\arguments{
\item{data}{An \linkS4class{MSnExp} object in \emph{centroid} mode.}

\item{expand_rt}{}
}
\description{
Function to apply peak picking parameters on a data subset
}
