% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Module_autotuner.R
\name{autotunerUI}
\alias{autotunerUI}
\title{Autotuner UI
`}
\usage{
autotunerUI(id)
}
\arguments{
\item{id}{character used to specify namespace, see \code{\link[shiny:NS]{shiny::NS}}}
}
\value{
a \code{shiny::\link[shiny]{tagList}} containing UI elements
}
\description{
Autotuner UI
`
}
