% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_stats_multi.R
\name{calculate_clustering}
\alias{calculate_clustering}
\title{Multivariate analysis - Clustering}
\usage{
calculate_clustering(
  abun_table,
  metadata,
  color_by,
  distance = "euclidean",
  cluster_algorithm = "ward.D2",
  add_kmeans = FALSE,
  k = 3,
  plot = TRUE,
  color_vector = NULL
)
}
\arguments{
\item{abun_table}{Feature abundance table}

\item{k}{Selecte K (for kmeans method)}

\item{method}{Clustering method}
}
\description{
Function to calculate ordination with different methods
}
