% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_stats_multi.R
\name{calculate_permanova}
\alias{calculate_permanova}
\title{Multivariate analysis - PERMANOVA}
\usage{
calculate_permanova(
  abun_table,
  metadata,
  vars,
  distance = "bray",
  assess = "model",
  use_interaction = FALSE,
  strata = NULL
)
}
\arguments{
\item{abun_table}{Feature abundance table with features as rows and samples as columns}

\item{vars}{Independent variables}

\item{distance}{Dissimilarity distance to calculate. Available options are
those from \code{vegan::vegdist()}}

\item{use_interaction}{Calculate interaction of selected variables}

\item{strata}{Variable to define sample blocks}
}
\value{
Object returned depend on the selected method
}
\description{
Function to perform a permutational analysis of variance
}
