% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_preprocessing.R
\name{centroid_check}
\alias{centroid_check}
\title{Check centroided}
\usage{
centroid_check(data, transform = TRUE)
}
\arguments{
\item{data}{An \linkS4class{MSnExp} object.}

\item{transform}{Transform the data into centroid mode}
}
\description{
This function checks if data is centroided
}
