% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_stats_univ.R
\name{check_model}
\alias{check_model}
\title{Check model performance}
\usage{
check_model(model_df)
}
\value{
The return value, if any, from executing the function.
}
\description{
Function to check the performance of the models that were fitted.
}
