% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UI_colorpicker.R
\name{colorPickerUI}
\alias{colorPickerUI}
\title{ColorPickerUI}
\usage{
colorPickerUI(id)
}
\arguments{
\item{id}{id for Namespacing}

\item{...}{UI elements to be passed to MetaboTandem app}
}
\value{
UI elements for colorpicker widget
}
\description{
UI for color picker
}
