% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_correspondence.R
\name{apply_correspondence}
\alias{apply_correspondence}
\alias{apply_correspondence.pd}
\alias{apply_correspondence.np}
\title{Apply Correspondence}
\usage{
apply_correspondence(data, metadata, method = c("pd", "np"), ...)

apply_correspondence.pd(
  data,
  metadata,
  group_by,
  bandwidth = 30,
  bin_size = 0.25,
  ppm = 0,
  min_fraction = 0.9,
  max_features = 30,
  ...
)

apply_correspondence.np(
  data,
  metadata,
  group_by,
  mzvsrtbal = 10,
  abs_mz = 0.2,
  abs_rt = 15,
  knn = 10,
  ...
)
}
\arguments{
\item{data}{An \linkS4class{MSnExp} object in \emph{centroid} mode.}

\item{metadata}{Sample information data.frame.}
}
\value{
MSnExp-class object after alignment.
}
\description{
Function to apply peak alignment
}
