% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{create_col_vector}
\alias{create_col_vector}
\title{Color vectors}
\usage{
create_col_vector(metadata, color_by = "AUTO")
}
\arguments{
\item{metadata}{Sample information file}

\item{color_by}{Column to color output}
}
\description{
Create color vectors
}
