% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_functions.R
\name{cycloess.norm}
\alias{cycloess.norm}
\title{Cyclic LOESS normalization}
\usage{
cycloess.norm(matrix)
}
\arguments{
\item{matrix}{Matrix or data frame with the unnormalized peak intensities
with peaks as rows and samples as columns}

\item{transform_data}{Logical value of whether log transform data or not}
}
\description{
Function to apply cyclic LOESS normalization to the data
}
