% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_data_filter.R
\name{apply_prevalence_filtering}
\alias{apply_prevalence_filtering}
\alias{apply_variance_filtering}
\alias{iqr_filtering}
\alias{sd_filtering}
\alias{mad_filtering}
\alias{rsd_filtering}
\title{Filter by prevalence}
\usage{
apply_prevalence_filtering(abun_table, min_perc_samples)

apply_variance_filtering(abun_table, filter_method, perc_remove)

iqr_filtering(abun_table, perc_remove)

sd_filtering(abun_table, perc_remove)

mad_filtering(abun_table, perc_remove)

rsd_filtering(abun_table, perc_remove)
}
\arguments{
\item{abun_table}{abun_table or data frame with the non-normalized peak intensities
with peaks as rows and samples as columns}

\item{min_perc_samples}{Minimum percentage of samples a feature need to be
present}

\item{filter_method}{Filter method to apply}

\item{perc_remove}{Percentage of features to remove from the analysis}
}
\description{
Function to filter features based on prevalence

Function to filter features with low variances
}
