% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Module_annotation.R
\name{dbAnnotationServer}
\alias{dbAnnotationServer}
\title{Database annotation server}
\usage{
dbAnnotationServer(id, data_proc, new_data = FALSE)
}
\arguments{
\item{id}{character used to specify namespace, see \code{\link[shiny:NS]{shiny::NS}}}

\item{ms1_data}{A \link{data.frame} with columns: 'name', 'mz' and 'rt'}
}
\description{
Database annotation server
}
