% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Module_annotation.R
\name{dbAnnotationUI}
\alias{dbAnnotationUI}
\title{Annotation UI
`}
\usage{
dbAnnotationUI(id)
}
\arguments{
\item{id}{character used to specify namespace, see \code{\link[shiny:NS]{shiny::NS}}}
}
\value{
a \code{shiny::\link[shiny]{tagList}} containing UI elements
}
\description{
Annotation UI
`
}
