% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_stats_univ.R
\name{differential_analysis}
\alias{differential_analysis}
\title{Differential analysis}
\usage{
differential_analysis(
  abun_table,
  metadata,
  norm_method,
  group,
  control_condition,
  treatment_condition
)
}
\value{
The return value, if any, from executing the function.
}
\description{
Function to perform differential abundance analysis among the specified groups
}
