% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Module_diffExpression.R
\name{diffExpressionUI}
\alias{diffExpressionUI}
\title{Ui for differential expression analysis}
\usage{
diffExpressionUI(id)
}
\arguments{
\item{id}{character used to specify namespace, see \code{\link[shiny:NS]{shiny::NS}}}
}
\value{
a \code{shiny::\link[shiny]{tagList}} containing UI elements
}
\description{
Ui for differential expression analysis
}
