% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Module_downloadResults.R
\name{download_ResultspreprocUI}
\alias{download_ResultspreprocUI}
\title{Download results (preprocessing) UI}
\usage{
download_ResultspreprocUI(id)
}
\arguments{
\item{id}{character used to specify namespace, see \code{\link[shiny:NS]{shiny::NS}}}
}
\value{
a \code{shiny::\link[shiny]{tagList}} containing UI elements
}
\description{
Download results (preprocessing) UI
}
