% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_extract.R
\name{extract_abundance_table}
\alias{extract_abundance_table}
\title{Extract abundance table}
\usage{
extract_abundance_table(data)
}
\arguments{
\item{data}{An \linkS4class{MSnExp} object in \emph{centroid} mode.}
}
\value{
The return value, if any, from executing the function.
}
\description{
Function to extract the abundance table
}
