% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_autotuner.R
\name{extract_autotuner}
\alias{extract_autotuner}
\title{Extract parameters AutoTuner}
\usage{
extract_autotuner(autotuner_obj, massThr)
}
\arguments{
\item{massThr}{Sample information data.frame.}
}
\description{
This function extracts the parameters estimated by AutoTuner
}
