% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_functions.R
\name{extract_feature_definition}
\alias{extract_feature_definition}
\title{Extract feature definitions}
\usage{
extract_feature_definition(data, feature_abundance, save_table = NULL)
}
\arguments{
\item{data}{An \linkS4class{MSnExp} object with peaks picked.}

\item{feature_abundance}{Data.frame with feature abundances per sample.}

\item{save_table}{Path to save the extracted table or \code{NULL}}
}
\description{
Function to apply peak picking parameters on a data subset
}
