% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_extract.R
\name{extract_feature_definitions}
\alias{extract_feature_definitions}
\title{Extract feature definitions}
\usage{
extract_feature_definitions(data)
}
\arguments{
\item{data}{An \linkS4class{MSnExp} object in \emph{centroid} mode.}
}
\value{
The return value, if any, from executing the function.
}
\description{
Function to extract the abundance table
}
