% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_extract.R
\name{extract_feature_spectra}
\alias{extract_feature_spectra}
\title{Extract feature spectra}
\usage{
extract_feature_spectra(data, rm_low_int = TRUE, min_peaks = 3)
}
\arguments{
\item{data}{An \linkS4class{MSnExp} object in \emph{centroid} mode.}
}
\value{
The return value, if any, from executing the function.

Define a function to remove low intensity peaks
}
\description{
Function to extract the abundance table
}
