% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_functions.R
\name{extract_MS2}
\alias{extract_MS2}
\title{Extract MS2 data}
\usage{
extract_MS2(data, save_spectra)
}
\arguments{
\item{data}{An \linkS4class{MSnExp} object with peaks picked.}

\item{save_table}{Path to save the extracted table or \code{NULL}}
}
\description{
Function to apply peak picking parameters on a data subset
}
