% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_extract.R
\name{extract_MS2_consensus}
\alias{extract_MS2_consensus}
\title{Extract MS2 data}
\usage{
extract_MS2_consensus(data)
}
\arguments{
\item{data}{An \linkS4class{MSnExp} object with peaks picked.}

\item{save_table}{Path to save the extracted table or \code{NULL}}
}
\description{
Function to apply peak picking parameters on a data subset
}
