% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_functions.R
\name{extract_number_peaks}
\alias{extract_number_peaks}
\title{Extract number of peaks}
\usage{
extract_number_peaks(data, save.table = FALSE, save.figure = FALSE)
}
\arguments{
\item{data}{An \linkS4class{MSnExp} object with peaks picked}
}
\description{
Function to apply peak picking parameters on a data subset
}
