% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_functions.R
\name{extract_spectra_table}
\alias{extract_spectra_table}
\title{Extract spectra table}
\usage{
extract_spectra_table(data, save_table = NULL)
}
\arguments{
\item{data}{An \linkS4class{MSnExp} object with peaks picked.}

\item{save_table}{Path to save the extracted table or \code{NULL}}
}
\description{
Function to apply peak picking parameters on a data subset
}
