% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_stats_univ.R
\name{fit_model}
\alias{fit_model}
\alias{fit_model.lm}
\alias{fit_model.lme}
\title{Fit linear models}
\usage{
fit_model(
  abun_table,
  metadata,
  model_type,
  vars = NULL,
  fix_vars = NULL,
  rand_vars = NULL
)

fit_model.lm(feature_table, vars)

fit_model.lme(feature_table, fix_vars, rand_vars)
}
\value{
The return value, if any, from executing the function.
}
\description{
Function to perform differential abundance analysis among the specified groups
}
