% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Module_gapFilling.R
\name{gapFillingUI}
\alias{gapFillingUI}
\title{Gap fill UI}
\usage{
gapFillingUI(id)
}
\arguments{
\item{id}{character used to specify namespace, see \code{\link[shiny:NS]{shiny::NS}}}
}
\value{
a \code{shiny::\link[shiny]{tagList}} containing UI elements
}
\description{
Gap fill UI
}
