% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_annotation.R
\name{get_annotation_tables}
\alias{get_annotation_tables}
\title{Get annotation tables}
\usage{
get_annotation_tables(
  feature_table,
  feature_spectra,
  selected_dbs = c("massbank", "mona", "hmdb_exp", "hmdb_pred", "gnps"),
  adducts = c("[M+H]+"),
  tolerance = 0.005,
  ppm = 5,
  req_precursor = TRUE,
  distance_thres = 0.5,
  candidates = 1
)
}
\arguments{
\item{feature_table}{Table with features mz and rt values}

\item{feature_spectra}{Features spectra}

\item{selected_dbs}{Vector with names of the databases to be used}

\item{adducts}{Vector with the adducts to be used for MS1 annotation}

\item{tolerance}{Allowed differences in m/z}

\item{ppm}{Allowed differences in ppm}

\item{req_precursor}{Check precursor for MS2 annotation}

\item{distance_thres}{Distance threshold for annotation}

\item{candidates}{How many annotations to keep for each feature}
}
\value{
List with annotations from each of the databases
}
\description{
Function to annotate features using MS1 and MS2 information
}
