% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_annotation.R
\name{get_molecular_class}
\alias{get_molecular_class}
\title{Classify metabolites}
\usage{
get_molecular_class(annotation_table)
}
\arguments{
\item{annotation_table}{Dataframe with all annotations}
}
\value{
Dataframe with molecular classification
}
\description{
Function to classify annotated metabolites using Classyfire
}
