% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_functions.R
\name{global.norm}
\alias{global.norm}
\title{Global are under the curve normalization}
\usage{
global.norm(matrix, transform_data = TRUE)
}
\arguments{
\item{matrix}{Matrix or data frame with the unnormalized peak intensities
with peaks as rows and samples as columns}

\item{transform_data}{Logical value of whether log transform data or not}
}
\description{
Function to apply global sum normalization to the data
}
