% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UI_home_app.R
\docType{data}
\name{home_ui}
\alias{home_ui}
\title{Homepage UI}
\format{
An object of class \code{shiny.tag.list} (inherits from \code{list}) of length 2.
}
\usage{
home_ui
}
\description{
UI for displaying MetaboTandem Homepage
}
\keyword{datasets}
