% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_load_data.R
\name{load_dataframe}
\alias{load_dataframe}
\title{Load dataframe}
\usage{
load_dataframe(file)
}
\arguments{
\item{file}{Path to the dataframe file}
}
\value{
A dataframe
}
\description{
Load dataframe
}
\details{
Load DataFrame

Function to load metadata into the MetaboTandem object
}
