% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Module_loadData.R
\name{load_dataServer}
\alias{load_dataServer}
\title{Load data server-side processing}
\usage{
load_dataServer(id)
}
\arguments{
\item{id}{character used to specify namespace,
see \code{shiny::\link[shiny]{NS}}\}
}
\value{
list with following components
\describe{
\item {data_cent}{A centroided \linkS4class{MSnExp} object}
\item {data_cent}{Dataframe with sample information}
}
}
\description{
Load data server-side processing
}
