% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Module_loadData.R
\name{load_dataUI}
\alias{load_dataUI}
\title{Load data UI}
\usage{
load_dataUI(id)
}
\arguments{
\item{id}{character used to specify namespace, see \code{\link[shiny:NS]{shiny::NS}}}
}
\value{
a \code{shiny::\link[shiny]{tagList}} containing UI elements
}
\description{
Load data UI
}
