% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_load_data.R
\name{load_metadata}
\alias{load_metadata}
\title{Load metadata}
\usage{
load_metadata(metadata_file)
}
\arguments{
\item{metadata_file}{Path to the metadata file}
}
\value{
Metadata table within the MetaboTandem object
}
\description{
Load metadata
}
\details{
Load metadata

Function to load metadata into the MetaboTandem object
}
