% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing_functions.R
\name{load_spectra_data}
\alias{load_spectra_data}
\title{Load Spectra data}
\usage{
load_spectra_data(datadir, metadata, format = "mzML")
}
\arguments{
\item{datadir}{Path to the directory with the data.}

\item{metadata}{Sample information data.frame.}

\item{format}{Format of the data.}
}
\description{
This function is intended to load the spectra data in mzML or mzXML format
using the xcms package.
}
