% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_stats.R
\name{max.norm}
\alias{max.norm}
\title{Max normalization}
\usage{
\method{max}{norm}(matrix, transform_data = TRUE)
}
\arguments{
\item{matrix}{Matrix or data frame with the unnormalized peak intensities
with peaks as rows and samples as columns}

\item{transform_data}{Logical value of whether log transform data or not}
}
\description{
Function to apply normalization based on most intense peak to the data
}
