% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_functions.R
\name{mean.norm}
\alias{mean.norm}
\title{Mean Normalzation
Function to apply global mean normalization to the data}
\usage{
\method{mean}{norm}(matrix, transform_data = TRUE)
}
\arguments{
\item{matrix}{Matrix or data frame with the unnormalized peak intensities
with peaks as rows and samples as columns}

\item{transform_data}{Logical value of whether log transform data or not}
}
\description{
Mean Normalzation
Function to apply global mean normalization to the data
}
