% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_stats.R
\name{median.norm}
\alias{median.norm}
\title{Median Normalization}
\usage{
\method{median}{norm}(matrix, transform_data = TRUE)
}
\arguments{
\item{matrix}{Matrix or data frame with the unnormalized peak intensities
with peaks as rows and samples as columns}

\item{transform_data}{Logical value of whether log transform data or not}
}
\description{
Function to apply global median normalization to the data
}
