% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_annotation.R
\name{merge_annotation_tables}
\alias{merge_annotation_tables}
\title{Merge annotation tables}
\usage{
merge_annotation_tables(annotation_tables_list, feature_table, candidates = 1)
}
\arguments{
\item{annotation_tables_list}{List with annotations to merge}

\item{candidates}{How many annotations to keep for each feature}
}
\value{
Dataframe with top annotation from all databases
}
\description{
Function to merge annotations from multiple databases
}
