% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_annotation_SIRIUS.R
\name{merge_SIRIUS_files}
\alias{merge_SIRIUS_files}
\title{Merge SIRIUS annotation}
\usage{
merge_SIRIUS_files(annot_dir)
}
\description{
Function to merge the annotation files retrieved using SIRIUS
}
