% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaboTandem-package.R
\docType{package}
\name{MetaboTandem-package}
\alias{MetaboTandem}
\alias{MetaboTandem-package}
\title{MetaboTandem}
\description{
Analysis and Visualization of LC-MS/MS data
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
