% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaboTandemApp.R
\name{MetaboTandemApp}
\alias{MetaboTandemApp}
\title{MetaboTandem UI}
\usage{
MetaboTandemApp()
}
\arguments{
\item{...}{UI elements to be passed to MetaboTandem app}

\item{id}{id for Namespacing}
}
\value{
UI elements for MetaboTandem
}
\description{
UI for MetaboTandem pipeline
}
