% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Module_loadData.R
\name{metadataServer}
\alias{metadataServer}
\title{Metadata uploading  server-side processing}
\usage{
metadataServer(id)
}
\arguments{
\item{id, }{character used to specify namespace, see \code{shiny::\link[shiny]{NS}}}

\item{input, }{output, session standard \code{shiny} boilerplate}
}
\value{
list with following components
\describe{
\item{xvar}{reactive character string indicating x variable selection}
\item{yvar}{reactive character string indicating y variable selection}
}
}
\description{
Metadata uploading  server-side processing
}
