% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Module_loadData.R
\name{metadataUI}
\alias{metadataUI}
\title{Metadata uploading UI}
\usage{
metadataUI(id)
}
\arguments{
\item{id, }{character used to specify namespace, see \code{shiny::\link[shiny]{NS}}}
}
\value{
a \code{shiny::\link[shiny]{tagList}} containing UI elements
}
\description{
Metadata uploading UI
}
