% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_functions.R
\name{mod_get_identification_table}
\alias{mod_get_identification_table}
\title{Modified - Get identification table}
\usage{
mod_get_identification_table(object, candidate.num = 3)
}
\arguments{
\item{object}{One or multiple metIdentifyClass objects.}

\item{candidate.num}{The number of candidates.}
}
\value{
A identification table \link{data.frame}.
}
\description{
Function to retrieve annotation table. This
function was modified from the package \href{https://github.com/tidymass/metid}{metid}
}
