% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_annotation.R
\name{mod_identify_all}
\alias{mod_identify_all}
\title{Modified - Identify metabolites using multiple databases}
\usage{
mod_identify_all(
  ms1.data,
  ms2.data,
  db_dir,
  parameter.list,
  process_ms2 = FALSE
)
}
\arguments{
\item{ms1.data}{A \link{data.frame} containing the columns 'FeatureID', 'mzmed' and 'rtmed'}

\item{ms2.data}{An \link{MSpectra} object containing MS2 data}

\item{db_dir}{Path where the databases are stored}

\item{parameter.list}{A list containing the parameters to process each database.
Parameters must be created using metid::metIdentifyParam()}
}
\value{
A list of \link{mzIdentifyClass} objects with the annotation results
}
\description{
Function to identify metabolites using multiple databases at once. This
function was modified from the package \href{https://github.com/tidymass/metid}{metid}
}
