% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Module_multivariate.R
\name{multivariateUI}
\alias{multivariateUI}
\title{Perform multivariate Analysis}
\usage{
multivariateUI(id)
}
\arguments{
\item{id}{character used to specify namespace, see \code{\link[shiny:NS]{shiny::NS}}}
}
\value{
a \code{shiny::\link[shiny]{tagList}} containing UI elements
}
\description{
Perform multivariate Analysis
}
