% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_functions.R
\name{nmds_ordination}
\alias{nmds_ordination}
\title{Function to calculate NMDS}
\usage{
nmds_ordination(abundance_matrix, metadata, mode = "ra", color_by)
}
\arguments{
\item{df}{Matrix or data frame with the unnormalized peak intensities
with peaks as rows and samples as columns}
}
\description{
Function to test all normalization methods and compare graphically
}
