% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_functions.R
\name{norm_methods_all}
\alias{norm_methods_all}
\title{Function to apply any normalization method required}
\usage{
norm_methods_all(df, norm_method, transform = TRUE)
}
\arguments{
\item{df}{Matrix or data frame with the unnormalized peak intensities
with peaks as rows and samples as columns}
}
\description{
Function to test all normalization methods and compare graphically
}
