% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_normalization.R
\name{apply_normalization}
\alias{apply_normalization}
\alias{global_norm}
\alias{median_norm}
\alias{mean_norm}
\alias{vsn_norm}
\alias{cycloess_norm}
\alias{max_norm}
\title{Normalize data}
\usage{
apply_normalization(abun_table, norm_method, log_transform = FALSE)

global_norm(abun_table, transform_data = TRUE)

median_norm(abun_table, transform_data = TRUE)

mean_norm(abun_table, transform_data = TRUE)

vsn_norm(abun_table)

cycloess_norm(abun_table)

max_norm(abun_table, transform_data = TRUE)
}
\arguments{
\item{abun_table}{Matrix or data frame with the unnormalized peak intensities
with peaks as rows and samples as columns}

\item{norm_method}{Normalization method to apply}

\item{transform_data}{Logical value of whether log transform data or not}
}
\description{
Function to apply different normalization methods to the data
}
