% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_stats.R
\name{normalize_by_all}
\alias{normalize_by_all}
\title{Test normalization methods}
\usage{
normalize_by_all(df)
}
\arguments{
\item{df}{Matrix or data frame with the unnormalized peak intensities
with peaks as rows and samples as columns}
}
\description{
Function to test all normalization methods and compare graphically
}
