% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_app.R
\name{old_MetaboTandemApp}
\alias{old_MetaboTandemApp}
\title{MetaboTandem UI}
\usage{
old_MetaboTandemApp()
}
\arguments{
\item{...}{UI elements to be passed to MetaboTandem app}

\item{id}{id for Namespacing}
}
\value{
UI elements for MetaboTandem
}
\description{
UI for MetaboTandem pipeline
}
