% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_stats_multi.R
\name{calculate_ordination}
\alias{calculate_ordination}
\alias{nmds_ordination}
\alias{pca_ordination}
\alias{pcoa_ordination}
\title{Multivariate analysis - Ordination}
\usage{
calculate_ordination(abun_table, method, distance = "bray", group = NULL)

nmds_ordination(abun_table, distance)

pca_ordination(abun_table)

pcoa_ordination(abun_table, distance)
}
\arguments{
\item{abun_table}{Feature abundance table}

\item{method}{Ordination method}

\item{distance}{Dissimilarity distance to calculate. Available options are
those from \code{vegan::vegdist()}}
}
\value{
Object returned depend on the selected method
}
\description{
Function to calculate ordination with different methods
}
