% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_stats_multi.R
\name{plot_ordination}
\alias{plot_ordination}
\alias{plot_ordination.nmds}
\alias{plot_ordination.pca}
\alias{plot_ordination.pcoa}
\title{Multivariate analysis - Plot Ordination}
\usage{
plot_ordination(
  ord_object,
  metadata,
  group_by,
  add_sample_names = FALSE,
  add_variables = FALSE,
  add_ellipse = FALSE,
  color_vector = NULL,
  abundances = NULL,
  plot = TRUE
)

plot_ordination.nmds(
  ord_object,
  metadata,
  group_by,
  add_sample_names,
  add_variables,
  add_ellipse,
  color_vector,
  abundances,
  plot
)

plot_ordination.pca(
  ord_object,
  metadata,
  group_by,
  add_sample_names,
  add_variables,
  add_ellipse,
  color_vector,
  plot
)

plot_ordination.pcoa(
  ord_object,
  metadata,
  group_by,
  add_sample_names,
  add_variables,
  add_ellipse,
  color_vector,
  abundances,
  plot
)
}
\arguments{
\item{ord_onj}{Feature abundance table}

\item{method}{Ordination method}

\item{distance}{Dissimilarity distance to calculate. Available options are
those from \code{vegan::vegdist()}}
}
\value{
Object returned depend on the selected method
}
\description{
Function to plot ordination results
}
