% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_annotation_SIRIUS.R
\name{parse_ms2_for_SIRIUS}
\alias{parse_ms2_for_SIRIUS}
\title{Function to parse ms2 data to work with SIRIUS}
\usage{
parse_ms2_for_SIRIUS(sp, TITLE = NULL)
}
\description{
Function to parse ms2 data to work with SIRIUS
}
