% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_peak_picking.R
\name{apply_peak_picking}
\alias{apply_peak_picking}
\alias{apply_peak_picking.cw}
\alias{apply_peak_picking.mq}
\alias{apply_peak_picking.mf}
\title{Apply Peak Picking}
\usage{
apply_peak_picking(data, method, ...)

apply_peak_picking.cw(
  data,
  ppm = 25,
  p_width = c(20, 50),
  snt = 3,
  noise = 1e+06,
  prefilter = c(1, 100),
  mz_diff = 0.001
)

apply_peak_picking.mq(
  data,
  ppm = 25,
  p_width = c(20, 50),
  snt = 3,
  noise = 1e+06,
  prefilter = c(1, 100),
  mz_diff = 0.001
)

apply_peak_picking.mf(
  data,
  bin = 0.1,
  fwhm = 30,
  sigma = 12.72,
  max = 10,
  steps = 2,
  cores = 1
)
}
\arguments{
\item{data}{An [MSnExp-class] object in *centroid* mode.}

\item{method}{Character. Specifies the peak picking method. Options include 'cw' for CentWave, 'mq' for Massifquant, and 'mf' for MatchedFilter.}

\item{ppm}{Numeric. Specifies the parts per million for m/z tolerance. Additional parameters include p_width, snt, noise, prefilter, mz_diff, bin, fwhm, sigma, max, and steps, each with specific roles in peak picking.}

\item{p_width}{Minium and maximum allowed peak width.}

\item{snt}{Minimum signal-to-noise threshold allowed.}

\item{noise}{Noise threshold.}

\item{prefilter}{Prefilter step cutoff (`c(k, I)`). Mass will be retained if they contain at least `k` peaks with intensity >= `I`.}

\item{BPPARAM}{BiocParallelParam-class object. Specifies the parallel processing parameters. Default is SnowParam with 4 workers.}
}
\value{
MSnExp-class object after applying peak picking.
}
\description{
Functions to perform peak picking on the data:

  `apply_peak_picking:` Wrapper to perform peak picking with any of the methods
}
\details{
Apply Peak Picking
Function to apply peak picking parameters on a data subset
}
