% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Module_peakPicking.R
\name{peakPickingServer}
\alias{peakPickingServer}
\title{Peak picking server-side processing}
\usage{
peakPickingServer(id, data)
}
\arguments{
\item{id}{character used to specify namespace, see \code{\link[shiny:NS]{shiny::NS}}}

\item{data}{dataframe with sample information}
}
\value{
\describe{
\item{data_cent_pp}{A \linkS4class{MSnExp} object with identified peaks}
}
}
\description{
Peak picking server-side processing
}
