% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_annotation.R
\name{plot_mirror}
\alias{plot_mirror}
\title{Plot spectra annotation mirror}
\usage{
plot_mirror(matches_list, merged_annotation_table, feature_id)
}
\arguments{
\item{matches_list}{List with matches to plot}
}
\value{
Dataframe with top annotation from all databases
}
\description{
Function to make a mirror plot of the annotation of a feature
}
