% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_stats_univ.R
\name{plot_model_sig_features}
\alias{plot_model_sig_features}
\title{Heatmap significant features}
\usage{
plot_model_sig_features(
  abundance_table,
  metadata,
  contrasts_results,
  color_by,
  cluster_feat = FALSE,
  cluster_samp = FALSE,
  color_vector = NULL
)
}
\value{
The return value, if any, from executing the function.
}
\description{
Function to check the performance of the models that were fitted.
}
