% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_stats_univ.R
\name{plot_volcano}
\alias{plot_volcano}
\title{Volcano plot}
\usage{
plot_volcano(diff_table, pval_thres, log2fc_thres, use_adjusted_pval = FALSE)
}
\value{
The return value, if any, from executing the function.
}
\description{
Function to perform differential abundance analysis among the specified groups
}
