% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ms2_functions.R
\name{read_mgf}
\alias{read_mgf}
\title{Read mgf data}
\usage{
read_mgf(file, separator = " ", threads = parallelly::availableCores())
}
\arguments{
\item{file}{Database file in mgf format}
}
\description{
This function is intended to read data in mgf format.
This function was modified from the \code{read_msp_mona()} function from
package \code{metID}
}
