% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rp_data.R
\docType{data}
\name{rp_data}
\alias{rp_data}
\title{Spectra data for RP LC-MS/MS experiment}
\format{
A \link{OnDiskMSnExp} object with two spectra
}
\usage{
rp_data
}
\description{
A dataset containing the trimmed spectra of two samples
from a reverse-phase (RP) LC-MS/MS experiment
}
\keyword{datasets}
