% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_annotation_SIRIUS.R
\name{separate_MS2}
\alias{separate_MS2}
\title{Separate MGF into groups}
\usage{
separate_MS2(ms2.data, outdir, n)
}
\description{
This function separates MS2 data into chunks to use SIRIUS web services
}
