% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Module_annotation_SIRIUS.R
\name{siriusAnnotationUI}
\alias{siriusAnnotationUI}
\title{Annotation using SIRIUS UI
`}
\usage{
siriusAnnotationUI(id)
}
\arguments{
\item{id}{character used to specify namespace, see \code{\link[shiny:NS]{shiny::NS}}}
}
\value{
a \code{shiny::\link[shiny]{tagList}} containing UI elements
}
\description{
Annotation using SIRIUS UI
`
}
