% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_autotuner.R
\name{start_autotuner}
\alias{start_autotuner}
\title{Start autotuner}
\usage{
start_autotuner(metadata, group, lag, threshold, influence, plot = FALSE)
}
\arguments{
\item{metadata}{Sample information data.frame.}

\item{group}{Grouping variable.}

\item{lag}{Chromatographic scan points to test if next point is significant}

\item{threshold}{How many times larger intensity need to be to
be considered significant}

\item{influence}{Factor to scale the magnitude of a significant scan}

\item{plot}{Logical value of whether to create a plot of the signals}
}
\description{
This function to create Autotuner object and get peak signals
}
