% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Module_statsSetup.R
\name{statsSetupUI}
\alias{statsSetupUI}
\title{Set options for statistical analysis}
\usage{
statsSetupUI(id)
}
\arguments{
\item{id}{character used to specify namespace, see \code{\link[shiny:NS]{shiny::NS}}}
}
\value{
a \code{shiny::\link[shiny]{tagList}} containing UI elements
}
\description{
Set options for statistical analysis
}
