% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_stats_univ.R
\name{test_contrasts}
\alias{test_contrasts}
\title{Test contrasts}
\usage{
test_contrasts(model_df, L)
}
\value{
The return value, if any, from executing the function.
}
\description{
Function to check the performance of the models that were fitted.
}
