% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing_functions.R
\name{test_peak_picking}
\alias{test_peak_picking}
\title{Test peak picking}
\usage{
test_peak_picking(
  data,
  mz.range,
  rt.range,
  p.width,
  snt,
  noise,
  prefilter = c(3, 100)
)
}
\arguments{
\item{data}{An \linkS4class{MSnExp} object in \emph{centroid} mode.}

\item{mz.range}{\emph{mz} range (\code{numeric(2)}) to test peak picking parameters.}

\item{rt.range}{\emph{rt} range (\code{numeric(2)}) to test peak picking parameters.}

\item{p.width}{Minium and maximum allowed peak width}

\item{snt}{Minimum signal-to-noise threshold allowed}

\item{noise}{Noise threshold}

\item{prefilter}{Prefilter step cutoff (\code{c(k, I)}). Mass will be retained if
they contain at least \code{k} peaks with intensity >= \code{I}}
}
\description{
Function to test peak picking parameters on a data subset
}
