% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UI_main.R
\name{ui_main}
\alias{ui_main}
\title{MetaboTandem UI}
\usage{
ui_main()
}
\arguments{
\item{...}{UI elements to be passed to MetaboTandem app}

\item{id}{id for Namespacing}
}
\value{
UI elements for MetaboTandem
}
\description{
UI for MetaboTandem pipeline
}
