% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_stats.R
\name{vsn.norm}
\alias{vsn.norm}
\title{Variance stabilization normalization
Function to apply variance stabilization normalization to the data}
\usage{
vsn.norm(matrix)
}
\arguments{
\item{matrix}{Matrix or data frame with the unnormalized peak intensities with
peaks as rows and samples as columns}

\item{transform_data}{Logical value of whether log transform data or not}
}
\description{
Variance stabilization normalization
Function to apply variance stabilization normalization to the data
}
