% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_gap_filling.R
\name{apply_gap_filling}
\alias{apply_gap_filling}
\title{Apply Gap Filling}
\usage{
apply_gap_filling(data, cores = 1)
}
\arguments{
\item{data}{An \linkS4class{MSnExp} object in \emph{centroid} mode.}
}
\value{
MSnExp-class object after alignment.
}
\description{
Function for gap filling
}
