% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_peak_picking.R
\name{apply_peak_refinement}
\alias{apply_peak_refinement}
\title{Apply Peak Refinement}
\usage{
apply_peak_refinement(
  data,
  expand_rt = 2,
  expand_mz = 0,
  ppm = 10,
  min_prop = 0.75
)
}
\arguments{
\item{data}{An [MSnExp-class] object in *centroid* mode.}

\item{expand_rt}{Numeric. Specifies the range to expand retention time during peak merging. Default is 2.}

\item{expand_mz}{Numeric. Specifies the range to expand m/z during peak merging. Default is 0.}

\item{ppm}{Numeric. Parts per million tolerance for peak merging.}

\item{min_prop}{Numeric. Minimum proportion of overlapping intensity required to merge peaks. Default is 0.75.}
}
\value{
MSnExp-class object after refining peaks by merging neighboring peaks.
}
\description{
Function to apply peak picking parameters on a data subset.
}
\details{
Apply Peak Refinement
Function to apply peak picking parameters on a data subset
}
