% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing_functions.R
\name{calculate_bpc}
\alias{calculate_bpc}
\title{Calculate BPC}
\usage{
calculate_bpc(
  data,
  metadata,
  show.plot = TRUE,
  save.figure = FALSE,
  color_by = "AUTO"
)
}
\arguments{
\item{data}{An \linkS4class{MSnExp} object.}

\item{metadata}{Sample information file.}

\item{show.plot}{Whether to show the plot.}

\item{save.figure}{Path to save the figure or \code{NULL}.}

\item{color_by}{Column to color output.}
}
\description{
Calculate and plot Base Peak Chromatogram
}
