% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing_functions.R
\name{calculate_tic}
\alias{calculate_tic}
\title{Calculate TIC}
\usage{
calculate_tic(
  data,
  metadata,
  show.plot = TRUE,
  save.figure = NULL,
  color_by = "AUTO"
)
}
\arguments{
\item{data}{An \linkS4class{MSnExp} object.}

\item{metadata}{Sample information file.}

\item{show.plot}{Wheter to show the plot.}

\item{save.figure}{Path to save the figure or \code{NULL}.}

\item{color_by}{Column to color output.}
}
\description{
Calculate and plot Total Ion Chromatogram
}
