% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Module_annotation.R
\name{dbAnnotationServer}
\alias{dbAnnotationServer}
\title{Database annotation server}
\usage{
dbAnnotationServer(id, ms1_data, ms2_data)
}
\arguments{
\item{id}{character used to specify namespace, see \code{\link[shiny:NS]{shiny::NS}}}

\item{data}{dataframe with sample information}
}
\value{
\describe{
\item{data_cent_pp}{A \linkS4class{MSnExp} object with identified peaks}
}
}
\description{
Database annotation server
}
