% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_functions.R
\name{extract_features}
\alias{extract_features}
\title{Extract features}
\usage{
extract_features(data, save_table = NULL)
}
\arguments{
\item{data}{An \linkS4class{MSnExp} object with peaks picked.}

\item{save_table}{Path to save the extracted table or \code{NULL}}

\item{metadata}{Sample information data.frame.}
}
\description{
Function to apply peak picking parameters on a data subset
}
