% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_extract.R
\name{extract_number_peaks}
\alias{extract_number_peaks}
\title{Extract number of peaks}
\usage{
extract_number_peaks(data, save_table = NULL, save_figure = NULL)
}
\arguments{
\item{data}{An \linkS4class{MSnExp} object with peaks picked}

\item{save_table}{Path to save the extracted table or \code{NULL}}

\item{save_figure}{Path to save the figure or \code{NULL}}
}
\description{
Function to apply peak picking parameters on a data subset
}
