% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_load_data.R
\name{load_metadata}
\alias{load_metadata}
\title{Load Metadata}
\usage{
load_metadata(metadata_file)
}
\arguments{
\item{metadata_file}{Path to the dataframe file}
}
\value{
A dataframe with metadata
}
\description{
Function will load metadata and check its format
}
\details{
Load Metadata

Function to load metadata
}
