% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_load_data.R
\name{load_spectra_data}
\alias{load_spectra_data}
\title{Load Spectra Data}
\usage{
load_spectra_data(datadir, metadata, format = "mzML", mode = "onDisk")
}
\arguments{
\item{datadir}{Path to the directory containing the data files.}

\item{metadata}{Contains sample information. Each row represents a sample and columns contain metadata information.}

\item{format}{Specifies the format of the data files. Default is 'mzML'. Supported formats include 'mzML' and 'mzXML'.}

\item{mode}{Specifies the mode of operation. Default is 'onDisk'. Other modes depend on xcms package capabilities.}
}
\value{
An object containing the loaded spectra data.
}
\description{
Loads spectra data in mzML or mzXML format using the xcms package.
}
\details{
Load Spectra data
This function is intended to load the spectra data in mzML or mzXML format
using the xcms package
}
