% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaboTandem-package.R
\docType{package}
\name{MetaboTandem-package}
\alias{MetaboTandem}
\alias{MetaboTandem-package}
\title{MetaboTandem}
\description{
Analysis and Visualization of LC-MS/MS data
}
\seealso{
Useful links:
\itemize{
  \item \url{https://coayala.github.io/MetaboTandem/}
}

}
\author{
\strong{Maintainer}: Christian Ayala-Ortiz \email{cayalaortiz@email.arizona.edu} (\href{https://orcid.org/0000-0001-8154-2378}{ORCID})

}
\keyword{internal}
