% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Module_loadData.R
\name{metadataServer}
\alias{metadataServer}
\title{Metadata uploading  server-side processing}
\usage{
metadataServer(id)
}
\arguments{
\item{id}{character used to specify namespace,
see \code{shiny::\link[shiny]{NS}}\}
}
\value{
A \link{data.frame} with sample information
}
\description{
Metadata uploading  server-side processing
}
