% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_peak_picking.R
\name{apply_peak_picking}
\alias{apply_peak_picking}
\alias{apply_peak_picking.cw}
\alias{apply_peak_picking.mq}
\alias{apply_peak_picking.mf}
\title{Apply Peak Picking}
\usage{
apply_peak_picking(data, method = c("cw", "mf", "mq"), cores = 1, ...)

apply_peak_picking.cw(
  data,
  BPPARAM,
  ppm = 25,
  p_width = c(20, 50),
  snt = 3,
  noise = 1e+06,
  prefilter = c(1, 100),
  mz_diff = 0.001,
  ...
)

apply_peak_picking.mq(
  data,
  BPPARAM,
  ppm = 25,
  p_width = c(20, 50),
  snt = 3,
  noise = 1e+06,
  prefilter = c(1, 100),
  mz_diff = 0.001,
  ...
)

apply_peak_picking.mf(
  data,
  BPPARAM,
  bin = 0.1,
  fwhm = 30,
  sigma = 12.72,
  max = 10,
  steps = 2,
  ...
)
}
\arguments{
\item{data}{An \linkS4class{MSnExp} object in \emph{centroid} mode.}

\item{method}{Character. Specifies the peak picking method. Options include 'cw' for CentWave, 'mq' for Massifquant, and 'mf' for MatchedFilter.}

\item{BPPARAM}{BiocParallelParam-class object. Specifies the parallel processing parameters. Default is SnowParam with 4 workers.}

\item{ppm}{Numeric. Specifies the parts per million for m/z tolerance. Additional parameters include p_width, snt, noise, prefilter, mz_diff, bin, fwhm, sigma, max, and steps, each with specific roles in peak picking.}

\item{p_width}{Minium and maximum allowed peak width.}

\item{snt}{Minimum signal-to-noise threshold allowed.}

\item{noise}{Noise threshold.}

\item{prefilter}{Prefilter step cutoff (\code{c(k, I)}). Mass will be retained if they contain at least \code{k} peaks with intensity >= \code{I}.}
}
\value{
MSnExp-class object after applying peak picking.
}
\description{
Functions to perform peak picking on the data:
\itemize{
\item \code{apply_peak_picking}
\itemize{
\item Wrapper to perform peak picking with any of the methods
}
\item \code{apply_peak_picking.cw}
\itemize{
\item Peak picking with the CentWave method
}
\item \code{apply_peak_picking.mf}
\itemize{
\item Peak picking with the Massifquant method
}
\item \code{apply_peak_picking.mq}
\itemize{
\item Peak picking with the MatchedFilter method
}
}
}
\details{
Apply Peak Picking
Function to apply peak picking parameters on a data subset
}
