% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_peak_picking.R
\name{test_peak_picking}
\alias{test_peak_picking}
\title{Test Peak Picking}
\usage{
test_peak_picking(
  data,
  mz_range,
  rt_range,
  p_width,
  snt,
  noise,
  prefilter = c(1, 100),
  cores = 1
)
}
\arguments{
\item{data}{An \linkS4class{MSnExp} object in \emph{centroid} mode.}

\item{snt}{Minimum signal-to-noise threshold allowed}

\item{noise}{Noise threshold}

\item{prefilter}{Prefilter step cutoff (\code{c(k, I)}). Mass will be retained if they contain at least \code{k} peaks with intensity >= \code{I}}

\item{...}{Parameters to pass to the selected method}

\item{mz.range}{\emph{mz} range (\code{numeric(2)}) to test peak picking parameters.}

\item{rt.range}{\emph{rt} range (\code{numeric(2)}) to test peak picking parameters.}

\item{p.width}{Minium and maximum allowed peak width}
}
\value{
Modified MSnExp-class object with tested peak picking parameters.
}
\description{
Tests peak picking parameters on a subset of the data.
}
\details{
Test peak picking
Function to test peak picking parameters on a data subset
}
